Add mobile menu toggling
tailwind config:
- Tried adding the screen sizes because for some reason, the navbar
would not center the menu items in desktop size. This was to no avail,
I modified the classes in the html directly, but there should be something
in here that I can tweak to get the desired results

menu-controls.js:
- Added some id's to the page because this is a small project, and I'm not
too concerned with scaling. I do wonder if I should produce HTML with jQuery,
particularly stuff that will repeat (e.g. the layout of each artist's profile)
Likewise, I'll have to add a generic event listener for every menu item, or something

- There is a weird race condition in the mobile menu click listener. Basically, both the
closing and opening button ("x" and hamburger icons respectively) run the risk of
both having the same class ("hidden") at a given time, or not having it at all, which
is undesirable. I have to find a way to write a lock for that (hopefully not so verbose)
I just didn't want to waste too much time thinking about this because of the deadline